Update text when needed properly
Previously the text wouldn't update if no one joined or left
a playing team.

This obviously was not optimal, since if the game started, the text
would not be removed.
Furthermore, if you went spec from RR, the text would remain.

This commit fixes those two issues by always notifying a client
if he changes team and always notifying all clients if the game state
changes.